// server/src/database/enums/question-theme.enum.ts
export enum QuestionTheme {
  Lyrics = 'lyrics',
  Charts = 'charts',
  Aesthetic = 'aesthetic',
  Timeline = 'timeline',
  Career = 'career',
  Audio = 'audio',
  Albums = 'albums',
  Songs = 'songs',
  Tours = 'tours',
  Trivia = 'trivia',
  Outfits = 'outfits',
  Mashups = 'mashups',
  Inspiration = 'inspiration',
  Mood = 'mood',
  Popularity = 'popularity',
  Tracklist = 'tracklist',
  Speed = 'speed',
  Visuals = 'visuals',
  Events = 'events'
}